gpu: ipu-v3: Fix BGR interface pix fmts, add RGB666
The interface pixel maps defined for BGR666 and LVDS666 in ipu-dc
were actually RGB order, fix them to be actual BGR order.

This means any users of the original BGR666 or LVDS666 mappings
were actually requesting RGB order mappings. So we need to add a
real RGB666 mapping to ipu-dc and then modify the users of "BGR666"
to ask for RGB666 instead. Those users are imx-ldb and imx-parallel
(there were no users of LVDS666).

Also modify the fsl-imx-drm.txt bindings docs to change the "bgr666"
supported interface format to "rgb666".

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>